+++ fabric/src/main/java/dev/qixils/crowdcontrol/plugin/fabric/interfaces/OriginalDisplayName.java
@@ -3,7 +3,7 @@ package dev.qixils.crowdcontrol.plugin.fabric.interfaces;
 import net.minecraft.network.chat.Component;
 import org.jetbrains.annotations.Nullable;
 
-public interface OriginalDisplayName extends dev.onyxstudios.cca.api.v3.component.Component {
+public interface OriginalDisplayName extends org.ladysnake.cca.api.v3.component.Component {
 
 	@Nullable Component getValue();