+++ fabric/src/main/java/dev/qixils/crowdcontrol/plugin/fabric/interfaces/ViewerMob.java
@@ -1,6 +1,6 @@
 package dev.qixils.crowdcontrol.plugin.fabric.interfaces;
 
-import dev.onyxstudios.cca.api.v3.component.Component;
+import org.ladysnake.cca.api.v3.component.Component;
 
 public interface ViewerMob extends Component {