+++ fabric/src/main/java/dev/qixils/crowdcontrol/plugin/fabric/interfaces/impl/ViewerMobImpl.java
@@ -1,6 +1,7 @@
 package dev.qixils.crowdcontrol.plugin.fabric.interfaces.impl;
 
 import dev.qixils.crowdcontrol.plugin.fabric.interfaces.ViewerMob;
+import net.minecraft.core.HolderLookup;
 import net.minecraft.nbt.CompoundTag;
 import org.jetbrains.annotations.NotNull;
 
@@ -18,12 +19,12 @@ public class ViewerMobImpl implements ViewerMob {
 	}
 
 	@Override
-	public void writeToNbt(@NotNull CompoundTag tag) {
+	public void writeToNbt(@NotNull CompoundTag tag, HolderLookup.@NotNull Provider registryLookup) {
 		tag.putBoolean("value", value);
 	}
 
 	@Override
-	public void readFromNbt(@NotNull CompoundTag tag) {
+	public void readFromNbt(@NotNull CompoundTag tag, HolderLookup.@NotNull Provider registryLookup) {
 		if (tag.contains("value")) {
 			value = tag.getBoolean("value");
 		}